PythonBE/src/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asteroid-filterbanks==0.4.0
attrs==25.1.0
audioread==3.0.1
av==14.2.0
beautifulsoup4
Brotli==1.1.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
ctranslate2==4.5.0
cycler==0.12.1
decorator==5.2.1
dnspython==2.7.0
docopt==0.6.2
einops==0.8.1
email_validator==2.2.0
fastapi
fastapi-cli==0.0.7
faster-whisper
filelock==3.17.0
firebase-admin
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-speech==2.31.0
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-generativeai
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.14.0
llvmlite==0.44.0
loguru
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
ngrok==1.4.0
numba==0.61.0
numpy==2.1.3
omegaconf==2.3.0
onnxruntime==1.21.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyannote.audio
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyngrok
pyparsing==3.2.1
pyphen==0.17.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
speechbrain==1.0.2
SpeechRecognition==3.14.1
SQLAlchemy==2.0.38
srt==3.5.3
starlette==0.46.0
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.6.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
weasyprint==64.1
webencodings==0.5.1
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.18.3
zopfli==0.2.3.post1